requirements.txt
PyYAML
python-Levenshtein
reCBZ
